+++ tests/end-to-end/application_api_test.go
@@ -3,8 +3,6 @@ package end_to_end
 import (
 	"context"
 	"fmt"
-	"os"
-	"reflect"
 	"testing"
 
 	"github.com/kyma-incubator/compass/components/director/pkg/graphql"
@@ -12,13 +10,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-var tc = testContext{graphqlizer: graphqlizer{}, gqlFieldsProvider: gqlFieldsProvider{}, cli: gcli.NewClient(getDirectorURL())}
-
-func (tc *testContext) RunQuery(ctx context.Context, req *gcli.Request, resp interface{}) error {
-	m := resultMapperFor(&resp)
-	return tc.cli.Run(ctx, req, &m)
-}
 
 func TestCreateApplicationWithAllSimpleFieldsProvided(t *testing.T) {
 	// GIVEN
@@ -749,227 +740,6 @@ func TestUpdateApplicationParts(t *testing.T) {
 	})
 }
 
-func TestRuntimeCreateUpdateAndDelete(t *testing.T) {
-	// GIVEN
-	ctx := context.Background()
-	givenInput := graphql.RuntimeInput{
-		Name:        "runtime-1",
-		Description: ptrString("runtime-1-description"),
-		Labels:      &graphql.Labels{"ggg": []string{"hhh"}},
-		Annotations: &graphql.Annotations{"kkk": "lll"},
-	}
-	runtimeInGQL, err := tc.graphqlizer.RuntimeInputToGQL(givenInput)
-	require.NoError(t, err)
-	actualRuntime := graphql.Runtime{}
-
-	// WHEN
-	createReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-			result: createRuntime(in: %s) {
-					%s
-				}
-			}`, runtimeInGQL, tc.gqlFieldsProvider.ForRuntime()))
-	saveQueryInExamples(t, createReq.Query(), "create runtime")
-	err = tc.RunQuery(ctx, createReq, &actualRuntime)
-
-	//THEN
-	require.NoError(t, err)
-	require.NotEmpty(t, actualRuntime.ID)
-	assert.Equal(t, givenInput.Name, actualRuntime.Name)
-	assert.Equal(t, *givenInput.Description, *actualRuntime.Description)
-	assert.Equal(t, *givenInput.Labels, actualRuntime.Labels)
-	assert.Equal(t, *givenInput.Annotations, actualRuntime.Annotations)
-	assert.NotNil(t, actualRuntime.AgentAuth)
-
-	// update runtime
-	givenInput.Description = ptrString("modified-runtime-1-description")
-	runtimeInGQL, err = tc.graphqlizer.RuntimeInputToGQL(givenInput)
-	require.NoError(t, err)
-
-	// WHEN
-	updateReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation{ 
-			result: updateRuntime(id: "%s", in: %s) {
-					%s
-				}
-			}`, actualRuntime.ID, runtimeInGQL, tc.gqlFieldsProvider.ForRuntime()))
-	saveQueryInExamples(t, updateReq.Query(), "update runtime")
-	err = tc.RunQuery(ctx, updateReq, &actualRuntime)
-
-	//THEN
-	require.NoError(t, err)
-	assert.Equal(t, *givenInput.Description, *actualRuntime.Description)
-
-	// add Label
-	actualLabel := graphql.Label{}
-
-	// WHEN
-	addLabelReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-			result: addRuntimeLabel(runtimeID: "%s", key: "%s", values: %s) {
-					%s
-				}
-			}`, actualRuntime.ID, "new-label", "[\"bbb\"]", tc.gqlFieldsProvider.ForLabel()))
-	err = tc.RunQuery(ctx, addLabelReq, &actualLabel)
-
-	//THEN
-	require.NoError(t, err)
-	assert.Equal(t, "new-label", actualLabel.Key)
-	assert.Len(t, actualLabel.Values, 1)
-	assert.Contains(t, actualLabel.Values, "bbb")
-
-	// add Annotation
-	actualAnnotation := graphql.Annotation{}
-
-	// WHEN
-	addAnnotationReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-			result: addRuntimeAnnotation(runtimeID: "%s", key: "%s", value: %s) {
-					%s
-				}
-			}`, actualRuntime.ID, "new-anno", `["zzz"]`, tc.gqlFieldsProvider.ForAnnotation()))
-	err = tc.RunQuery(ctx, addAnnotationReq, &actualAnnotation)
-
-	//THEN
-	require.NoError(t, err)
-	assert.Equal(t, "new-anno", actualAnnotation.Key)
-	assert.Equal(t, []interface{}{"zzz"}, actualAnnotation.Value)
-
-	// get runtime and validate runtimes and annotations
-	getRuntimeReq := gcli.NewRequest(
-		fmt.Sprintf(`query {
-			result: runtime(id: "%s") {
-					%s
-				}
-			}`, actualRuntime.ID, tc.gqlFieldsProvider.ForRuntime()))
-	err = tc.RunQuery(ctx, getRuntimeReq, &actualRuntime)
-	require.NoError(t, err)
-	assert.Len(t, actualRuntime.Labels, 2)
-	assert.Len(t, actualRuntime.Annotations, 2)
-
-	// delete label
-
-	// WHEN
-	delLabelReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-				result: deleteRuntimeLabel(runtimeID: "%s", key: "%s") {
-						%s
-					}
-				}
-		`, actualRuntime.ID, "new-label", tc.gqlFieldsProvider.ForLabel()))
-	err = tc.RunQuery(ctx, delLabelReq, nil)
-
-	//THEN
-	require.NoError(t, err)
-	// delete annotation
-
-	// WHEN
-	delAnnoReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-				result: deleteRuntimeAnnotation(runtimeID: "%s", key: "%s") {
-						%s
-				}
-			}`, actualRuntime.ID, "new-anno", tc.gqlFieldsProvider.ForAnnotation()))
-	err = tc.RunQuery(ctx, delAnnoReq, nil)
-
-	//THEN
-	require.NoError(t, err)
-
-	// delete runtime
-
-	// WHEN
-	delReq := gcli.NewRequest(fmt.Sprintf(`mutation{result: deleteRuntime(id: "%s") {%s}}`, actualRuntime.ID, tc.gqlFieldsProvider.ForRuntime()))
-	saveQueryInExamples(t, delReq.Query(), "delete runtime")
-	err = tc.RunQuery(ctx, delReq, nil)
-
-	//THEN
-	require.NoError(t, err)
-}
-
-func TestSetAndDeleteAPIAuth(t *testing.T) {
-	// GIVEN
-	// create application
-	ctx := context.Background()
-	placeholder := "app"
-	in := generateSampleApplicationInput(placeholder)
-
-	appInputGQL, err := tc.graphqlizer.ApplicationInputToGQL(in)
-	require.NoError(t, err)
-	createReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-  				result: createApplication(in: %s) {
-    					%s
-					}
-				}`, appInputGQL, tc.gqlFieldsProvider.ForApplication()))
-	actualApp := ApplicationExt{}
-	err = tc.RunQuery(ctx, createReq, &actualApp)
-	require.NoError(t, err)
-	require.NotEmpty(t, actualApp.ID)
-	defer deleteApplication(t, actualApp.ID)
-
-	// create runtime
-	runtimeInput := graphql.RuntimeInput{
-		Name:        "runtime-1",
-		Description: ptrString("runtime-1-description"),
-		Labels:      &graphql.Labels{"ggg": []string{"hhh"}},
-		Annotations: &graphql.Annotations{"kkk": "lll"},
-	}
-	runtimeInGQL, err := tc.graphqlizer.RuntimeInputToGQL(runtimeInput)
-	require.NoError(t, err)
-	actualRuntime := graphql.Runtime{}
-	createRuntimeReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-				result: createRuntime(in: %s) {
-						%s
-					}
-				}`, runtimeInGQL, tc.gqlFieldsProvider.ForRuntime()))
-	err = tc.RunQuery(ctx, createRuntimeReq, &actualRuntime)
-	require.NoError(t, err)
-	require.NotEmpty(t, actualRuntime.ID)
-
-	defer deleteRuntime(t, actualRuntime.ID)
-
-	actualRuntimeAuth := graphql.RuntimeAuth{}
-
-	// WHEN
-	// set Auth
-	authIn := graphql.AuthInput{
-		Credential: &graphql.CredentialDataInput{
-			Basic: &graphql.BasicCredentialDataInput{
-				Username: "x-men",
-				Password: "secret",
-			}}}
-
-	authInStr, err := tc.graphqlizer.AuthInputToGQL(&authIn)
-	require.NoError(t, err)
-	setAuthReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-			result: setAPIAuth(apiID: "%s", runtimeID: "%s", in: %s) {
-					%s
-				}
-			}`, actualApp.Apis.Data[0].ID, actualRuntime.ID, authInStr, tc.gqlFieldsProvider.ForRuntimeAuth()))
-	err = tc.RunQuery(ctx, setAuthReq, &actualRuntimeAuth)
-
-	//THEN
-	require.NoError(t, err)
-	require.NotNil(t, actualRuntimeAuth.Auth)
-	assert.Equal(t, actualRuntime.ID, actualRuntimeAuth.RuntimeID)
-	actualBasic, ok := actualRuntimeAuth.Auth.Credential.(*graphql.BasicCredentialData)
-	require.True(t, ok)
-	assert.Equal(t, "x-men", actualBasic.Username)
-	assert.Equal(t, "secret", actualBasic.Password)
-	// delete Auth
-	delAuthReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-			result: deleteAPIAuth(apiID: "%s",runtimeID: "%s") {
-					%s
-				} 
-			}`, actualApp.Apis.Data[0].ID, actualRuntime.ID, tc.gqlFieldsProvider.ForRuntimeAuth()))
-	err = tc.RunQuery(ctx, delAuthReq, nil)
-	require.NoError(t, err)
-
-}
-
 func TestQueryApplications(t *testing.T) {
 	// GIVEN
 	idsToRemove := make([]string, 3)
@@ -1055,94 +825,6 @@ func TestQuerySpecificApplication(t *testing.T) {
 	assert.Equal(t, createdID, actualApp.ID)
 }
 
-func TestQueryRuntimes(t *testing.T) {
-	// GIVEN
-	ctx := context.Background()
-
-	idsToRemove := make([]string, 3)
-	defer func() {
-		for _, id := range idsToRemove {
-			if id != "" {
-				deleteRuntime(t, id)
-			}
-		}
-	}()
-
-	for i := 0; i < 3; i++ {
-		givenInput := graphql.RuntimeInput{
-			Name: fmt.Sprintf("runtime-%d", i),
-		}
-		runtimeInGQL, err := tc.graphqlizer.RuntimeInputToGQL(givenInput)
-		require.NoError(t, err)
-		createReq := gcli.NewRequest(
-			fmt.Sprintf(`mutation {
-				result: createRuntime(in: %s) {
-						%s
-					} 
-				}`, runtimeInGQL, tc.gqlFieldsProvider.ForRuntime()))
-		actualRuntime := graphql.Runtime{}
-		err = tc.RunQuery(ctx, createReq, &actualRuntime)
-		require.NoError(t, err)
-		require.NotEmpty(t, actualRuntime.ID)
-		idsToRemove[i] = actualRuntime.ID
-	}
-	actualPage := graphql.RuntimePage{}
-
-	// WHEN
-	queryReq := gcli.NewRequest(
-		fmt.Sprintf(`query {
-			result: runtimes {
-					%s
-				}
-			}`, tc.gqlFieldsProvider.Page(tc.gqlFieldsProvider.ForRuntime())))
-	err := tc.RunQuery(ctx, queryReq, &actualPage)
-	saveQueryInExamples(t, queryReq.Query(), "query runtimes")
-
-	//THEN
-	require.NoError(t, err)
-	assert.Len(t, actualPage.Data, 3)
-	assert.Equal(t, 3, actualPage.TotalCount)
-
-}
-
-func TestQuerySpecificRuntime(t *testing.T) {
-	// GIVEN
-	ctx := context.Background()
-	givenInput := graphql.RuntimeInput{
-		Name: "runtime-1",
-	}
-	runtimeInGQL, err := tc.graphqlizer.RuntimeInputToGQL(givenInput)
-	require.NoError(t, err)
-	createReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation {
-			result: createRuntime(in: %s) {
-					%s
-				}
-			}`, runtimeInGQL, tc.gqlFieldsProvider.ForRuntime()))
-	createdRuntime := graphql.Runtime{}
-	err = tc.RunQuery(ctx, createReq, &createdRuntime)
-	require.NoError(t, err)
-	require.NotEmpty(t, createdRuntime.ID)
-
-	defer deleteRuntime(t, createdRuntime.ID)
-	queriedRuntime := graphql.Runtime{}
-
-	// WHEN
-	queryReq := gcli.NewRequest(
-		fmt.Sprintf(`query {
-			result: runtime(id: "%s") {
-					%s
-				}
-			}`, createdRuntime.ID, tc.gqlFieldsProvider.ForRuntime()))
-	err = tc.RunQuery(ctx, queryReq, &queriedRuntime)
-	saveQueryInExamples(t, queryReq.Query(), "query specific runtime")
-
-	//THEN
-	require.NoError(t, err)
-	assert.Equal(t, createdRuntime.ID, queriedRuntime.ID)
-	assert.Equal(t, createdRuntime.Name, queriedRuntime.Name)
-}
-
 func TestTenantSeparation(t *testing.T) {
 	//TODO
 }
@@ -1193,16 +875,6 @@ func deleteApplication(t *testing.T, id string) {
 	require.NoError(t, tc.RunQuery(context.Background(), req, nil))
 }
 
-func deleteRuntime(t *testing.T, id string) {
-	delReq := gcli.NewRequest(
-		fmt.Sprintf(`mutation{deleteRuntime(id: "%s") {
-				id
-			}
-		}`, id))
-	err := tc.RunQuery(context.Background(), delReq, nil)
-	require.NoError(t, err)
-}
-
 func fixBasicAuth() *graphql.AuthInput {
 	return &graphql.AuthInput{
 		Credential: &graphql.CredentialDataInput{
@@ -1241,47 +913,4 @@ func fixDepracatedVersion1() *graphql.VersionInput {
 		ForRemoval:      ptrBool(true),
 		DeprecatedSince: ptrString("v5"),
 	}
-}
-
-// testContext contains dependencies that help executing tests
-type testContext struct {
-	graphqlizer       graphqlizer
-	gqlFieldsProvider gqlFieldsProvider
-	cli               *gcli.Client
-}
-
-func getDirectorURL() string {
-	url := os.Getenv("DIRECTOR_GRAPHQL_API")
-	if url == "" {
-		url = "http://127.0.0.1:3000/graphql"
-	}
-	return url
-}
-
-// resultMapperFor returns generic object that can be passed to Run method for storing response.
-// In GraphQL, set `result` alias for your query
-func resultMapperFor(target interface{}) genericGQLResponse {
-	if reflect.ValueOf(target).Kind() != reflect.Ptr {
-		panic("target has to be a pointer")
-	}
-	return genericGQLResponse{
-		Result: target,
-	}
-}
-
-type genericGQLResponse struct {
-	Result interface{} `json:"result"`
-}
-
-func saveQueryInExamples(t *testing.T, query string, exampleName string) {
-	// TODO uncomment this in https://github.com/kyma-incubator/compass/issues/77
-	//t.Helper()
-	//sanitizedName := strings.Replace(exampleName, " ", "-", -1)
-	//// replace uuids with constant value
-	//r, err := regexp.Compile("[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}")
-	//require.NoError(t, err)
-	//query = r.ReplaceAllString(query, "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa")
-	//content := fmt.Sprintf("# This example is generated from test compass_e2e_test.go\n%s", query)
-	//err = ioutil.WriteFile(fmt.Sprintf("%s/src/github.com/kyma-incubator/compass/components/director/examples/%s.graphql", os.Getenv("GOPATH"), sanitizedName), []byte(content), 0660)
-	//require.NoError(t, err)
 }
